backend/requirements.txt
Flask
Flask-CORS
ultralytics
opencv-python
Pillow
numpy
